Fix the related links and topics blocks when referenced topic is deleted
Fix DRUP-794.

Adds as guard check around the get topics method in relatedLinks to check the
loaded term is an instance of TermInterface (can be null if the topic is
deleted), Also applies simmilar fix to ReleatedTopicsBlock as they both
try to use the term object after loading it directly without checking it
exists.