Add Service Path where not yet included by pathauto for nodes opt-ing into service hiearchy.
Related:
 * localgovdrupal/localgov_services#31
 * localgovdrupal/localgov#63
 * localgovdrupal/localgov_guides#2

Rather than maintain (enforce?) pathauto for all content types that include
the services hiearchy field, add it in if not present.

It's much easier to maintain that handling actual pathauto entitie (and the
now quite complex conditions). It doesn't prevent users from overriding or
changing these. The most clear downside to doing it on the pattern hook on
saving the entity is it is a little more magic where it is coming from.

This also continues to the pattern of treating the field name
'localgov_services_parent' as the identifier of an entity that can be
opting into services.