Simple test to verify ChildRelationshipsUI does not crash bad term
If the taxonomy term assigned to a child node does not exist, it should be
skipped over, not try to fetch a taxonomy term label.

This test is just a simple call for the formAlter method of
EntityChildRelationshipUi, which in turn calls childrenField which is where
the problem of loading a non existent term causing a WSOD was located.

Without the check for if ($topic->entity instanceof TermInterface) {
this test will fail.

This test could also be expanded in the future to check that
$form['localgov_services_navigation_children'] contains the correct
details about the elements refernecing the parent node.