Add search summary to search highlight result displahy mode
Fix #172

Add the search highlight pseduo field to search result display mode for all
service content types.

Note: This will mean that whilst not a hard dependency, we are essentially
requiring search_api and only supporting search api for search with this
display mode, else it will be an empty display.