Update all services blocks to use node contexts
Set a node context to all service blocks and update the block base and all service
block functions to use the contextValue('node') if present.

Update each of the provided blocks to set a context_mapping for the current node route
Provide update hooks for CTA, releated links and related topics blocks
Provide a backward compatible way of accessing the current node via $this->node
for blocks that have not updated to use a node context so they can still get the current
node. (todo: decide if this should then be deprecated).