Remove the unused $build theme page_header
Fix #289

Remove the old page_header from the $build array of the service status page
controller. This was a left over way of adding the page header from BHCC
Drupal 8 / Miggle site and was never used in Localgov Drupal.
It's presence was causing a Theme hook page_header not found warning in
watchdog.